@@ -17,7 +17,7 @@
 	<description>GovPay Modulo per la creazione delle stampe</description>
 
 	<properties>
-		<java.version>11</java.version>
+		<java.version>21</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<maven.compiler.release>${java.version}</maven.compiler.release>
@@ -520,7 +528,7 @@
 							<goal>xjc</goal>
 						</goals>
 						<configuration>
-							<target>2.1</target>
+							<target>3.0</target>
 							<sources>
 								<source>src/main/resources/xsd/govpayStampe.xsd</source>
 							</sources>
@@ -534,7 +542,7 @@
 							<goal>xjc</goal>
 						</goals>
 						<configuration>
-							<target>2.1</target>
+							<target>3.0</target>
 							<sources>
 								<source>src/main/resources/xsd/avvisiPagamento_v2.xsd</source>
 							</sources>
