@@ -13,7 +13,7 @@ public class JasperConfig {
 	public Jaxb2Marshaller jaxb2MarshallerAvvisoPagamento() {
 		Jaxb2Marshaller marshaller = new Jaxb2Marshaller();
 		marshaller.setPackagesToScan("it.govpay.stampe.model.v1");
-		marshaller.setMarshallerProperties(Collections.singletonMap(javax.xml.bind.Marshaller.JAXB_FRAGMENT, true));
+		marshaller.setMarshallerProperties(Collections.singletonMap(jakarta.xml.bind.Marshaller.JAXB_FRAGMENT, true));
 		return marshaller;
 	}
 	
@@ -21,7 +21,7 @@ public class JasperConfig {
 	public Jaxb2Marshaller jaxb2MarshallerAvvisoPagamentoBilingue() {
 		Jaxb2Marshaller marshaller = new Jaxb2Marshaller();
 		marshaller.setPackagesToScan("it.govpay.stampe.model.v2");
-		marshaller.setMarshallerProperties(Collections.singletonMap(javax.xml.bind.Marshaller.JAXB_FRAGMENT, true));
+		marshaller.setMarshallerProperties(Collections.singletonMap(jakarta.xml.bind.Marshaller.JAXB_FRAGMENT, true));
 		return marshaller;
 	}
 }
