@@ -1,6 +1,6 @@
 package it.govpay.stampe.controller;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
