@@ -2,7 +2,7 @@ package it.govpay.stampe.validator;
 
 import java.util.List;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 import org.springframework.stereotype.Component;
 
