@@ -85,7 +85,7 @@ class UC_3_AvvisoStandardFailTest {
         assertNotNull(problem.getString("detail"));
         assertEquals(400, problem.getInt("status"));
         assertEquals("Bad Request", problem.getString("title"));
-        assertTrue(problem.getString("detail").contains("Content type 'text/html' not supported"));
+        assertTrue(problem.getString("detail").contains("Content-Type 'text/html' is not supported"));
         assertEquals("https://www.rfc-editor.org/rfc/rfc9110.html#name-400-bad-request", problem.getString("type"));
 		
 	}
